ENH: Adding engine parameter to Series.map
